+++ Common Core/philosophers/philo/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: pjelinek <pjelinek@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/31 07:30:05 by netrunner         #+#    #+#             */
-/*   Updated: 2025/09/25 15:00:11 by pjelinek         ###   ########.fr       */
+/*   Updated: 2025/09/25 16:09:23 by pjelinek         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,9 +54,9 @@ typedef struct s_data
 {
 	int64_t		start_time_ms;
 	int64_t		timestamp;
-	int64_t		time_to_die;
-	int64_t		time_to_eat;
-	int64_t		time_to_sleep;
+	int			time_to_die;
+	int			time_to_eat;
+	int			time_to_sleep;
 	int			number_of_meals;
 	int			number_of_philos;
 	int			count;
@@ -67,25 +67,25 @@ typedef struct s_data
 	t_mutex		mutex;
 }	t_data;
 
-long		ft_atoi(const char *str);
+long long	ft_atoi(const char *str);
 void		*ft_calloc(size_t nmemb, size_t size);
 bool		input_check(char **av);
 void		print_lock(char *str, t_data *data);
 void		seperate_philos(t_data *data);
 
-void	cleanup(t_data *data);
-void	mutex_cleanup(pthread_mutex_t *mtx, int i, t_data *data);
-void	mutex_destroy(t_data *data);
-bool	mutex_init(t_data *data);
-void	thread_cleanup(pthread_t *thread, int i);
+void		cleanup(t_data *data);
+void		mutex_cleanup(pthread_mutex_t *mtx, int i, t_data *data);
+void		mutex_destroy(t_data *data);
+bool		mutex_init(t_data *data);
+void		thread_cleanup(pthread_t *thread, int i);
 
-void	set_forks(t_data *data);
-void	set_philo(t_data *data);
-bool	start_threads(t_data *data);
+void		set_forks(t_data *data);
+void		set_philo(t_data *data);
+bool		start_threads(t_data *data);
 
-void	set_starttime(t_data *data);
-int64_t	gettime(void);
-void	print_timestamp(t_data *data);
+void		set_starttime(t_data *data);
+int64_t		gettime(void);
+void		print_timestamp(t_data *data);
